[refactor/#46] Move study group update logic from MyPageService to GroupManagementService git status
